% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_values}
\alias{check_values}
\title{Single values conform to list}
\usage{
check_values(value)
}
\arguments{
\item{value}{A list of relays}
}
\value{
True if valid
}
\description{
If a single string value is provided, convert it to a list.
Else, return the original list.
}
