% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{config_env}
\alias{config_env}
\title{Setting environment to store configuration details for the session}
\format{
An object of class \code{environment} of length 0.
}
\usage{
config_env
}
\description{
Setting environment to store configuration details for the session
}
\keyword{datasets}
