% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_user_notes.R
\name{fetch_user_notes}
\alias{fetch_user_notes}
\title{Notes from a user}
\usage{
fetch_user_notes(
  authors = NA,
  event_refs = NA,
  pubkey_refs = NA,
  since = NA,
  until = NA,
  limit = 2000
)
}
\arguments{
\item{authors}{List or string of npub or HEX formatted author(s)}

\item{event_refs}{List or string of event references}

\item{pubkey_refs}{List or string of pub key references}

\item{since}{INT of interval start}

\item{until}{INT of interval termination}

\item{limit}{INT of #notes to fetch per relay (Defaults to 2000)}
}
\value{
Result stored in json format
}
\description{
Looks up notes posted by a user.
}
