% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{send_post_request}
\alias{send_post_request}
\title{Util function for sending post requests}
\usage{
send_post_request(url, body)
}
\arguments{
\item{body}{The body of the post request}
}
\value{
True if successful.
}
\description{
Util function for sending post requests
}
