% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_user_dm.R
\name{send_user_dm}
\alias{send_user_dm}
\title{Send a direct note to another user}
\usage{
send_user_dm(text, recipient_public_key)
}
\arguments{
\item{text}{The message to post to the note.}

\item{recipient_public_key}{The public id associated with the user.}
}
\description{
Allows the user to send a direct message to another nostr id
}
