% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_note.R
\name{send_user_note}
\alias{send_user_note}
\title{Send a note}
\usage{
send_user_note(text)
}
\arguments{
\item{note}{The message to post to the note.}
}
\value{
True if succesful
}
\description{
Allows the owner of a nostr public id to publish a note.
}
