% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{verify_api_config_success}
\alias{verify_api_config_success}
\title{Verify api config succesful}
\usage{
verify_api_config_success()
}
\value{
True if successful.
}
\description{
Verify api config succesful
}
