% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_user_notes.R
\name{fetch_user_notes}
\alias{fetch_user_notes}
\title{Notes from a user}
\usage{
fetch_user_notes(
  authors = NA,
  event_refs = NA,
  pubkey_refs = NA,
  since = NA,
  until = NA,
  limit = NA
)
}
\arguments{
\item{authors}{A list of the nostr npub's or hex's of an author.}
}
\value{
Result stored in json format
}
\description{
Looks up notes posted by a user.
}
